refactor: Rename package from comand to command (typo)
