+++ src/main/java/com/oadultradeepfield/smartotter/command/AddTodoCommand.java
@@ -1,4 +1,4 @@
-package com.oadultradeepfield.smartotter.comand;
+package com.oadultradeepfield.smartotter.command;
 
 import com.oadultradeepfield.smartotter.task.ToDoTask;
