+++ src/main/java/com/oadultradeepfield/smartotter/command/CommandContext.java
@@ -1,4 +1,4 @@
-package com.oadultradeepfield.smartotter.comand;
+package com.oadultradeepfield.smartotter.command;
 
 import com.oadultradeepfield.smartotter.task.Task;
 import java.util.List;
