+++ src/main/java/com/oadultradeepfield/smartotter/command/DeleteCommand.java
@@ -1,30 +1,30 @@
-package com.oadultradeepfield.smartotter.comand;
+package com.oadultradeepfield.smartotter.command;
 
 import com.oadultradeepfield.smartotter.SmartOtterException;
 import com.oadultradeepfield.smartotter.task.Task;
 import com.oadultradeepfield.smartotter.util.CustomIO;
 import java.util.List;
 
-/** Marks a specified task as done. */
-public record MarkCommand(int taskNumber) implements Executable {
+/** Deletes the specified task from the list. */
+public record DeleteCommand(int taskNumber) implements Executable {
   /**
-   * Creates a {@code MarkCommand} for the given task number.
+   * Creates a {@code DeleteCommand} for the given task number.
    *
-   * @param taskNumber the 1-based index of the task to mark as done
+   * @param taskNumber the 1-based index of the task to delete as done
    */
-  public MarkCommand {}
+  public DeleteCommand {}
 
   /**
-   * Parses input to create a {@code MarkCommand}.
+   * Parses input to create a {@code DeleteCommand}.
    *
    * @param taskNumber string representing the 1-based task number
-   * @return a new {@code MarkCommand} instance
+   * @return a new {@code DeleteCommand} instance
    * @throws SmartOtterException if the input is not a valid integer
    */
   public static Executable fromInput(String taskNumber) throws SmartOtterException {
     try {
       int num = Integer.parseInt(taskNumber);
-      return new MarkCommand(num);
+      return new DeleteCommand(num);
     } catch (NumberFormatException e) {
       throw new SmartOtterException("Invalid task number: %s".formatted(taskNumber));
     }
@@ -36,8 +36,15 @@ public record MarkCommand(int taskNumber) implements Executable {
     List<Task> tasks = context.tasks();
     if (taskNumber >= 1 && taskNumber <= tasks.size()) {
       Task task = tasks.get(taskNumber - 1);
-      task.setDone(true);
-      CustomIO.printPretty("Great work! I have marked the task as done!\n%s".formatted(task));
+      tasks.remove(task);
+
+      String message =
+          """
+                        Got it! I have deleted the task:
+                            %s
+                        Now you have %d tasks left."""
+              .formatted(task, context.tasks().size());
+      CustomIO.printPretty(message);
     } else {
       CustomIO.printPretty(
           CustomIO.formatError("There is no task with number %d".formatted(taskNumber)));
