+++ src/main/java/com/oadultradeepfield/smartotter/command/Executable.java
@@ -1,4 +1,4 @@
-package com.oadultradeepfield.smartotter.comand;
+package com.oadultradeepfield.smartotter.command;
 
 import com.oadultradeepfield.smartotter.SmartOtterException;
